avoid circular import and replace print()
